Add i2cdataclient for ipc and unit tests
Added i2cdataclient for ipc to our c++ program. The two are very coupled
together but I can't think of a good way to decouple that dependency.
Also added unit tests for it which hopefully will prove as a basic
example for adding unit tests and mocking out hardware requirements in
the future.