Add communication class for use with Xbee
Added the communication class for use with the Xbee module. Currently
this just sends a json version of whatever data it is given, and
receives a json object and returns the loaded version of that json
object. Some work should be done to make it non-blocking. Some work
should also be done to handle bad transmissions and lost transmissions.
Some basic unit tests confirm functionality of the receive and send
functions.

Added pytest to run all unit tests. This allows us to stop using the
tedious structure of the builtin unittest and provides more options for
parametizing tests in the future. Updated the CI scripts to reflect
this.

Added setup.py to allow the package to be installed as edittable via
pip. This allows us to ditch those annoying path injection lines in the
tests and examples and is a more permanent solution.

Separated tests based on whether they did or did not use the dronekit
simulator. The simulator takes forever to run tests and I found it to be
too slow to make a quick change and have to wait 40 seconds to get the
unit test results. Now you can get results more quickly and just run the
simulator tests every once in a while.

Added flake8 to the CI script. I'd like to use pylint or make flake8 a
little more aggressive in its checks, but I don't feel like configuring
pylint to chill out a bit or flake8 to get meaner.